fix(json): remove shadowed org.json dep and stop passing Float to JSONObject
Real crash root cause (reported by user):

  NoSuchMethodError: No Virtual method
  put(Ljava/lang/String;F)Lorg/json/JSONObject;
  in class Lorg/json/JSONObject;
  @HttpUploader.kt:206

The module declared org.json:json:20240303, so Kotlin resolved
JSONObject.put(String, float) at compile time. At runtime Android
always loads the platform org.json from core-libart.jar (bootclasspath
beats the app classloader), which has no float overload. Result: an
Error (not Exception) that escaped every catch(Exception) and silently
killed the process - which is why the app died without any tap, on the
15-minute periodic location upload path.

Fixes:
- drop org.json:json (platform already ships org.json)
- HttpUploader: accuracy_m -> accuracyM.toDouble()
- LocalFileUploader: speech_confidence -> confidence.toDouble()

Both were the only Float-typed JSONObject.put args in the repo.

+++ mobile/src/main/java/com/omi4wos/mobile/storage/HttpUploader.kt
@@ -203,7 +203,7 @@ class HttpUploader(
             val json = JSONObject()
                 .put("lat", lat)
                 .put("lon", lon)
-                .put("accuracy_m", accuracyM)
+                .put("accuracy_m", accuracyM.toDouble())  // 必须是 double：平台 org.json 无 put(String,float)
                 .put("source", source)
                 .put("timestamp", sampledAtMs)
             val body = json.toString()
